refactor: streamline oracle event handling and message encoding
 - Removed `BlockchainEventRouter` and its interface, simplifying the event routing mechanism.
 - Introduced `BlockchainEventHandler` as an abstract base for handling blockchain events, improving modularity.
 - Updated `NativeTokenMinter` to inherit from `BlockchainEventHandler`, aligning with the new event handling structure.
 - Refactored message encoding in `PortalMessage` to use a 16-byte nonce, reducing overhead from 52 to 36 bytes.
 - Consolidated `sendMessage` and `sendMessageWithData` into a single `send` function in `GravityPortal`, enhancing API clarity.
 - Adjusted tests to reflect changes in message encoding and event handling, ensuring comprehensive coverage.

+++ src/oracle/evm/native_token_bridge/IGTokenBridge.sol
@@ -16,7 +16,7 @@ interface IGTokenBridge {
     /// @param recipient The recipient address on Gravity
     /// @param amount The amount of tokens locked
     /// @param nonce The portal nonce for this bridge operation
-    event TokensLocked(address indexed from, address indexed recipient, uint256 amount, uint256 indexed nonce);
+    event TokensLocked(address indexed from, address indexed recipient, uint256 amount, uint128 indexed nonce);
 
     // ========================================================================
     // ERRORS
@@ -45,7 +45,7 @@ interface IGTokenBridge {
     function bridgeToGravity(
         uint256 amount,
         address recipient
-    ) external payable returns (uint256 messageNonce);
+    ) external payable returns (uint128 messageNonce);
 
     // ========================================================================
     // VIEW FUNCTIONS
@@ -69,4 +69,3 @@ interface IGTokenBridge {
         address recipient
     ) external view returns (uint256 requiredFee);
 }
-
